requirements.txt
torch>=2.4
torchvision>=0.19
tqdm>=4.66
PyYAML>=6.0
numpy>=2.0
pandas>=2.2
matplotlib>=3.8
